Add base entity to Switch as X
